Add a basic test for sending events over SSL to a RELP server, it also move the RelpClient to the support classes within the specs folder, as it belongs there for testing.
