+++ spec/spec_helper.rb
@@ -1,6 +1,7 @@
 # encoding: utf-8
 require "logstash/devutils/rspec/spec_helper"
 require "socket"
+require "support/client"
 
 module RelpHelpers
